Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/frontend/requirements.txt
Flask==2.2.2
requests==2.28.1
pylint==2.15.5
gunicorn==20.1.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.21.0
opentelemetry-instrumentation-flask==0.35b0
opentelemetry-instrumentation-requests==0.35b0
opentelemetry-instrumentation==0.35b0
opentelemetry-sdk==1.14.0
prometheus-client==0.15.0
opentelemetry-api==1.14.0